pyproject.toml
[project]
name = "advanced-data-vis-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=2.18.2",
    "dash-bootstrap-components>=1.7.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "ipykernel>=6.29.5",
    "geopandas>=1.0.1",
    "openpyxl>=3.1.5",
    "sqlalchemy>=2.0.40",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
